Migrate install-server-plugin off client/out/
install-server-plugin.mjs required() compiled output under
client/out/, an edge invisible to tsc/ESLint/git grep that a stale
build yields wrong behavior for rather than an error. Split it into
a thin launcher (stubs vscode via Module._load, then dynamically
imports the real script) and installServerPluginMain.ts, which
imports client/src directly — type-checked and staleness-proof.

Add tsx to run the .ts entry point directly, client/tsconfig.bin.json
to type-check it without emitting into client/out/, and a
test:server:install-plugin script wiring both together for local runs and CI.

Delete testConnectionImports.test.ts: its source-text import scan
existed only to guard the compiled-output edge, which no longer
exists.
