+++ package.json
@@ -2259,8 +2259,9 @@
   ],
   "scripts": {
     "vscode:prepublish": "npm run bundle",
-    "compile": "npm run compile:client && npm run compile:server && npm run compile:mcp",
+    "compile": "npm run compile:client && npm run compile:bin && npm run compile:server && npm run compile:mcp",
     "compile:client": "tsc -p client/tsconfig.json",
+    "compile:bin": "tsc -p client/tsconfig.bin.json",
     "compile:server": "tsc -p server/tsconfig.json",
     "compile:mcp": "tsc -p mcp-server/tsconfig.json",
     "bundle": "node esbuild.mjs",
@@ -2282,6 +2283,7 @@
     "test:server:start": "npm run test:server:start --workspace client",
     "test:server:stop": "npm run test:server:stop --workspace client",
     "test:server:list": "npm run test:server:list --workspace client",
+    "test:server:install-plugin": "tsx client/bin/install-server-plugin.mjs",
     "test:server": "npm test --workspace server",
     "test:client": "npm test --workspace client",
     "test:mcp": "npm test --workspace mcp-server",
@@ -2313,6 +2315,7 @@
     "ovsx": "^0.10.5",
     "playwright-bdd": "^9.2.0",
     "prettier": "^3.9.5",
+    "tsx": "^4.23.1",
     "typescript": "^5.7.2",
     "typescript-eslint": "^8.20.0",
     "vscode-oniguruma": "^2.0.1",
